requirements.txt
huggingface-hub==0.23.2
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
pymysql==1.1.1
speechbrain==0.5.16
torch==2.2.2
torchaudio==2.2.2
numpy==1.26.4
scipy==1.12.0
SpeechRecognition==3.10.3
soundfile==0.12.1
